initialize

+++ src/main/java/com/mapita/mapita_backend/service/impl/ProductServiceImpl.java
@@ -2,9 +2,8 @@ package com.mapita.mapita_backend.service.impl;
 
 import com.mapita.mapita_backend.dto.ProductDto;
 import com.mapita.mapita_backend.entity.ProductEntity;
-import com.mapita.mapita_backend.entity.UserEntity;
 import com.mapita.mapita_backend.mapper.ProductMapper;
-import com.mapita.mapita_backend.repository.CompanyRepository;
+import com.mapita.mapita_backend.repository.OrderRepository;
 import com.mapita.mapita_backend.repository.ProductRepository;
 import com.mapita.mapita_backend.service.ProductService;
 import lombok.RequiredArgsConstructor;
@@ -18,13 +17,11 @@ import java.util.stream.Collectors;
 public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository repository;
-    private final CompanyRepository companyRepository;
     private final ProductMapper mapper;
+    private final OrderRepository orderRepository;
     @Override
     public ProductDto save(ProductDto dto) {
         ProductEntity entity = mapper.toEntity(dto);
-        entity.setCompany(companyRepository.findById(dto.getCompanyId())
-                .orElseThrow(() -> new RuntimeException("company not found!")));
         return mapper.toDto(repository.save(entity));
     }
 
@@ -40,8 +37,6 @@ public class ProductServiceImpl implements ProductService {
         ProductEntity existProduct = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found"));
         existProduct.setProductName(dto.getProductName());
-        existProduct.setCompany(companyRepository.findById(dto.getCompanyId())
-                .orElseThrow(() -> new RuntimeException("Company not found")));
         return mapper.toDto(repository.save(existProduct));
     }
 
@@ -54,6 +49,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void delete(Long id) {
+        long count = orderRepository.countActiveOrFutureOrders(id);
+        if (count > 0) {
+            throw new RuntimeException("Cannot delete product with active");
+        }
         repository.deleteById(id);
     }
 }
